requirements.txt
fastapi
uvicorn[standard]
celery
redis
psycopg2-binary
requests
beautifulsoup4
playwright
python-dotenv
